fix: use window title from option instead of shell
The problem is when we set our shell to change window title with this
command (or something similar):
```
print -Pn "\e]0;%~\a"
```

If we use `st -t something`, instead of using `something` for window
title, `st` will use title from the shell.

In my opinion, this should not be an expected behavior. When i set
window title from command line, i don't want anything to change its
window title.